+++ Runtime/AssemblyInfo.cs
@@ -3,4 +3,5 @@ using System.Runtime.CompilerServices;
 using WhiteSparrow.Shared.Logging;
 
 [assembly: InternalsVisibleTo("WhiteSparrow.Logging.Editor")]
+[assembly: InternalsVisibleTo("WhiteSparrow.Logging.Tests")]
 [assembly: AssemblyVersion(Chirp.Version)]